Add new field per person to host modules/tutorials
Previously, the tag system was used to identify who is in which module, but it has proven to be annying when it
comes to searching because it is difficult to differentiate between modules and tutorials,
as well as which tutorial is connected to which module using it.
This commit adds a new field to Person comprising of module and tutorial objects to remedy the problem.