Fix bug for delete module command
When deleting a module, related tutorials in the GUI as well as those
in the students' modules & tutorials fields should be deleted as well.
The description in the UG has been updated to reflect this. I have
also modified the length requirement for addModule and ensured that
the module output on the GUI is always capitalised.